extend all handle values to 64bit
the eed entity handle needs 8 byte. so the handle.value and absref would too.

+++ src/hash.h
@@ -1,7 +1,7 @@
 /*****************************************************************************/
 /*  LibreDWG - free implementation of the DWG file format                    */
 /*                                                                           */
-/*  Copyright (C) 2018-2019 Free Software Foundation, Inc.                   */
+/*  Copyright (C) 2018-2019, 2023 Free Software Foundation, Inc.             */
 /*                                                                           */
 /*  This library is free software, licensed under the terms of the GNU       */
 /*  General Public License as published by the Free Software Foundation,     */
@@ -15,8 +15,8 @@
 
 /*
  * hash.h: simple fast int hashmap for the object_ref map,
- *         mapping uint32_t to uint32_t.
- *         0 kys and values are disallowed even if there's no deletion.
+ *         mapping uint64_t to uint64_t.
+ *         0 keys and values are disallowed even if there's no deletion.
  * written by Reini Urban
  */
 
@@ -25,23 +25,23 @@
 #include <inttypes.h>
 
 #define HASH_LOAD 75 // in percent. recommended is 50
-#define HASH_NOT_FOUND (uint32_t) - 1
+#define HASH_NOT_FOUND (uint64_t)-1
 
 struct _hashbucket
 {
-  uint32_t key;
-  uint32_t value;
+  uint64_t key;
+  uint64_t value;
 };
 typedef struct _inthash
 {
   struct _hashbucket *array; /* of key, value pairs */
-  uint32_t size;
-  uint32_t elems; // to get the fill rate
+  uint64_t size;
+  uint64_t elems; // to get the fill rate
 } dwg_inthash;
 
-dwg_inthash *hash_new (uint32_t size);
-uint32_t hash_get (dwg_inthash *hash, uint32_t key);
-void hash_set (dwg_inthash *hash, uint32_t key, uint32_t value);
+dwg_inthash *hash_new (uint64_t size);
+uint64_t hash_get (dwg_inthash *hash, uint64_t key);
+void hash_set (dwg_inthash *hash, uint64_t key, uint64_t value);
 void hash_free (dwg_inthash *hash);
 
 #endif
